backend/requirements.txt
FastAPI
uvicorn
torch
torchvision
Pillow
pandas
ultralytics
sqlalchemy
passlib